% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_event_and_group_information.R
\name{add_event_and_group_information}
\alias{add_event_and_group_information}
\title{Add Event Information and Group Information to Megaplot Data}
\usage{
add_event_and_group_information(
  data,
  summary_stats,
  event1,
  event2,
  event3,
  event4,
  updated_event1,
  updated_event2,
  updated_event3,
  updated_event4,
  data_selection
)
}
\arguments{
\item{data}{list with megaplot data}

\item{summary_stats}{list with megaplot data summary}

\item{event1}{character with first event variable}

\item{event2}{character with second event variable}

\item{event3}{character with third event variable}

\item{event4}{character with fourth event variable}

\item{updated_event1}{character with uploaded event 1}

\item{updated_event2}{character with uploaded event 2}

\item{updated_event3}{character with uploaded event 3}

\item{updated_event4}{character with uploaded event 4}

\item{data_selection}{character of data upload method ("Use demo data"/"Data upload")}
}
\description{
Add Event Information and Group Information to Megaplot Data
}
