% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sorting_information.R
\name{add_sorting_information}
\alias{add_sorting_information}
\title{Add Sorting Information to Data in Megaplots}
\usage{
add_sorting_information(
  data_frame,
  select_subsetting,
  randet,
  seed,
  nshow,
  start,
  specific_ids,
  event_levels,
  select_sorting,
  select_grouping,
  select_events
)
}
\arguments{
\item{data_frame}{list with megaplot data}

\item{select_subsetting}{Select subsets}

\item{randet}{character value ('deterministic'/'random')}

\item{seed}{integer with seed number for random selection}

\item{nshow}{integer with number of subjects displayed}

\item{start}{integer with row selection for subsetting}

\item{specific_ids}{character vector with specific subject identifier which should be displayed}

\item{event_levels}{character vector with event levels}

\item{select_sorting}{character with sorting variable selection}

\item{select_grouping}{character vector with grouping variable(s) selection}

\item{select_events}{character vector with event variable selection}
}
\description{
Add Sorting Information to Data in Megaplots
}
