% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers_create_file.R
\name{calc_days_with}
\alias{calc_days_with}
\title{Helper function to calculate days with event and days with event group}
\usage{
calc_days_with(data, calc_event_group = TRUE, calc_event = TRUE)
}
\arguments{
\item{data}{A data frame containing event data with columns for subject ID,
event group, event, event start time, and event end time.}

\item{calc_event_group}{Logical indicating whether to calculate days for event groups. Default is TRUE.}

\item{calc_event}{Logical indicating whether to calculate days for individual events. Default is TRUE.}
}
\value{
A data frame with the number of days associated with each subject's
events and event groups.
}
\description{
This function calculates the number of days with specific events
for each subject in the provided dataset.
Expects columns subjectid, event_group, event, event_time,
event_time_end.
}
