% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers_create_file.R
\name{calc_time_to_first}
\alias{calc_time_to_first}
\title{Helper function to calculate time to first event and time to first event group}
\usage{
calc_time_to_first(data, calc_event_group = TRUE, calc_event = TRUE)
}
\arguments{
\item{data}{A data frame containing event data with columns for subject ID,
event group, event, event start time, and event end time.}

\item{calc_event_group}{Logical indicating whether to calculate time to the first
event group. Default is TRUE.}

\item{calc_event}{Logical indicating whether to calculate time to the first
individual event. Default is TRUE.}
}
\value{
A data frame with the time to the first occurrence of each subject's
events and event groups.
}
\description{
This function calculates the time to the first occurrence of specific events
and event groups for each subject in the provided dataset.
Expects columns subjectid, event_group, event, event_time,
event_time_end.
}
