% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createfile_events.R
\name{createFile.events}
\alias{createFile.events}
\title{Function to create event level megaplots dataset from ADaM datasets}
\usage{
createFile.events(
  mp_data,
  path_data,
  id = "USUBJID",
  data_filter = NULL,
  param = list(c("AEBODSYS", "AEDECOD"), c("AEBODSYS", "AELLT"), c("AEDECOD", "AESEV"),
    c("AEDECOD", "AESER")),
  prefix = NULL,
  event_start = c("ASTDT", "AESTDT", "ADY"),
  event_end = c("AENDT", "AEENDT"),
  calc_time_to_first = FALSE
)
}
\arguments{
\item{event_end}{}
}
\description{
Function to create event level megaplots dataset from ADaM datasets
}
