% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createfile_old_to_new.R
\name{createfile.old_to_new}
\alias{createfile.old_to_new}
\title{Function to create megaplots file for rebuild (v??? and later) from prepared datasets for old version (v??? and earlier)}
\usage{
createfile.old_to_new(
  path_data,
  path_data_b = NULL,
  subjectid = "subjectid",
  event_time
)
}
\arguments{
\item{path_data}{Path to the subject-level dataset}

\item{path_data_b}{Path to the events dataset}

\item{subjectid}{Name of the subject ID column (as a string)}

\item{event_time}{Name of the event time column (as a string)}
}
\description{
Function to create megaplots file for rebuild (v??? and later) from prepared datasets for old version (v??? and earlier)
}
