% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createfile_sl.R
\name{createFile.sl}
\alias{createFile.sl}
\title{Function to create subject level megaplots dataset from ADSL (ADaM) dataset}
\usage{
createFile.sl(
  path_adsl,
  id = "USUBJID",
  data_filter = NULL,
  display_start_date = c("REFSTDT", "RFSTDT", "RFSTDTC", "RFICDT", "RANDDT", "TRTSTDT"),
  display_end_date = c("REFENDT", "RFENDTC", "RFENDT", "LVDT", "WDICDT"),
  relative_day_1 = c("TRTSTDT", "TRTSDT"),
  trt = TRUE,
  trtstdt = c("TRTSTDT", "TRTSDT"),
  trtendt = c("TRTENDT", "TRTEDT")
)
}
\arguments{
\item{path_adsl}{Path to the adsl-dataset}

\item{id}{unique subject identifier.}

\item{data_filter}{Subset dataset accoring to the filter conditions. Conditions should be wrapped in '' and concatenated by c()}

\item{trtendt}{}
}
\description{
Function to create subject level megaplots dataset from ADSL (ADaM) dataset
}
