% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_event_summary.R
\name{draw_event_summary}
\alias{draw_event_summary}
\title{Title}
\usage{
draw_event_summary(
  megaplot_prepared_data = megaplot_prepared_data(),
  megaplot_filtered_data = megaplot_filtered_data(),
  select_grouping = NULL,
  event_summary_cutoff = 1
)
}
\arguments{
\item{select_grouping}{character vector with grouping variables}

\item{event_summary_cutoff}{numeric value used as cutoff for hover labels displayed}
}
\description{
Title
}
