% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_kaplan_meier.R
\name{draw_kaplan_meier}
\alias{draw_kaplan_meier}
\title{Title}
\usage{
draw_kaplan_meier(
  megaplot_prepared_data = megaplot_prepared_data(),
  megaplot_filtered_data = megaplot_filtered_data(),
  select_grouping = NULL,
  select_event_kaplan_meier,
  select_strata_var = NULL
)
}
\arguments{
\item{select_strata_var}{}
}
\description{
Title
}
