% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mega_plot.R
\name{draw_mega_plot}
\alias{draw_mega_plot}
\title{Draws a 'Megaplot' graph}
\usage{
draw_mega_plot(
  megaplot_prepared_data = megaplot_prepared_data(),
  megaplot_filtered_data = megaplot_filtered_data(),
  select_grouping = NULL,
  line_width = 3,
  line_width_subjects,
  event_tooltips = TRUE,
  switch_legend_grouping = TRUE,
  sort_event_groups,
  sequencing_object,
  sequencing_switch,
  reference_line_1,
  reference_line_2,
  reference_line_3,
  reference_line_1_value,
  reference_line_2_value,
  reference_line_3_value,
  reference_line_1_value2,
  reference_line_2_value2,
  reference_line_3_value2,
  reference_line_1_color,
  reference_line_2_color,
  reference_line_3_color
)
}
\arguments{
\item{megaplot_prepared_data}{data.frame with subject information used for the subject lines}

\item{megaplot_filtered_data}{data.frame with selected events for the event lines}

\item{select_grouping}{character vector with grouping variable names}

\item{line_width}{numeric value for event line width}

\item{line_width_subjects}{numeric value for subject line width}

\item{event_tooltips}{logical value if event tooltips should be turned on/off}

\item{switch_legend_grouping}{logical value if events should be grouped in}

\item{sort_event_groups}{character vector with drag and drop order of event groups}

\item{sequencing_object}{seriation object created with TraMineR package}

\item{sequencing_switch}{logical value of sequencing should be turned on/off}

\item{reference_line_1}{logical value if first reference line/rect should be drawn}

\item{reference_line_2}{logical value if second reference line/rect should be drawn}

\item{reference_line_3}{logical value if third reference line/rect should be drawn}

\item{reference_line_1_value}{numeric value with start value for first reference line/rect (x-axis)}

\item{reference_line_2_value}{numeric value with start value for second reference line/rect (x-axis)}

\item{reference_line_3_value}{numeric value with start value for third reference line/rect (x-axis)}

\item{reference_line_1_value2}{numeric value with end value for first reference line/rect (x-axis)}

\item{reference_line_2_value2}{numeric value with end value for second reference line/rect (x-axis)}

\item{reference_line_3_value2}{numeric value with end value for third reference line/rect (x-axis)}

\item{reference_line_1_color}{character with hex color code for first reference line/rect}

\item{reference_line_2_color}{character with hex color code for second reference line/rect}

\item{reference_line_3_color}{character with hex color code for third reference line/rect
plotly legend (default: TRUE)}
}
\description{
Draws a 'Megaplot' graph
}
