% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_megaplot.R
\name{draw_megaplot}
\alias{draw_megaplot}
\title{Draw Megaplot main graph}
\usage{
draw_megaplot(
  megaplot_data,
  select_color,
  par_settings,
  background_stripes,
  background_stripes_length,
  event_levels,
  xlim,
  ylim,
  lines_instead_symbols,
  line_width,
  lines_options,
  y_axis_label,
  reference_line_1,
  reference_line_1_value,
  reference_line_2,
  reference_line_2_value,
  reference_line_3,
  reference_line_3_value,
  select_events,
  color_subject_line_by_first_event
)
}
\arguments{
\item{megaplot_data}{list with megaplot data A and B}

\item{select_color}{character vector with color definition ("plot.bg","plot.lines","plot.wp","plot.id","axleg.bg","cont.bg")}

\item{par_settings}{list with par settings ("mar","grLab")}

\item{background_stripes}{logical value if background stripes should be drawn}

\item{background_stripes_length}{integer value of background stripes length}

\item{event_levels}{character vector with event levels}

\item{xlim}{numeric vector with x-axis range}

\item{ylim}{numeric vector with y-axis range}

\item{lines_instead_symbols}{logical if lines should be displayed for events}

\item{line_width}{numeric value for line width for subject time lines}

\item{lines_options}{character for line option (two options available: Adjecent or Overlaying)}

\item{y_axis_label}{character for y axis label}

\item{reference_line_1}{logical if  vertical reference line should be displayed}

\item{reference_line_1_value}{integer where the vertical reference line should be displayed}

\item{reference_line_2}{logical if  vertical reference line should be displayed}

\item{reference_line_2_value}{integer where the vertical reference line should be displayed}

\item{reference_line_3}{logical if  vertical reference line should be displayed}

\item{reference_line_3_value}{integer where the vertical reference line should be displayed}

\item{select_events}{character vector with event names}

\item{color_subject_line_by_first_event}{logical if only first event should be displayed as colored line}
}
\description{
Draw Megaplot main graph
}
