% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_megaplot_legend.R
\name{draw_megaplot_legend}
\alias{draw_megaplot_legend}
\title{Draw Megaplot Legend}
\usage{
draw_megaplot_legend(megaplot_data, select_color)
}
\arguments{
\item{megaplot_data}{list with megaplot data A and B}

\item{select_color}{character vector with color definition ("plot.bg","plot.lines","plot.wp","plot.id","axleg.bg","cont.bg")}
}
\description{
Draw Megaplot Legend
}
