% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_help_text_dropdown_button.R
\name{help_text_dropdown_button}
\alias{help_text_dropdown_button}
\title{helper function for helptext in megaplots application}
\usage{
help_text_dropdown_button(id, text, right)
}
\arguments{
\item{id}{character with identifier name}

\item{text}{character string with help text}

\item{right}{logical (TRUE/FALSE) if help text should appear on right side}
}
\description{
helper function for helptext in megaplots application
}
