% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mp_init.R
\name{init_mp_object}
\alias{init_mp_object}
\title{Initialize an empty Megaplots data builder}
\usage{
init_mp_object()
}
\value{
An object of class `mp_data_builder`.
}
\description{
Returned object is meant to be used with [add.sl_data()], [add.events()],
and [finalize_mp_data()]. "events" starts as an empty table with the core
Megaplots event columns so rows can be stacked with [add.events()].
}
