% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_megaplots.R
\name{launch_megaplots}
\alias{launch_megaplots}
\title{launch_megaplots - Launches the Megaplots application}
\usage{
launch_megaplots(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\value{
A shiny app
}
\description{
Starts the Megaplots application in the client's browser.
}
\examples{
\dontrun{
launch_megaplots()
}

}
\keyword{megaplots}
