% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megaplots_demo_data.R
\docType{data}
\name{megaplots_demo_data}
\alias{megaplots_demo_data}
\title{List of datasets A and B containing example data}
\format{
List with data frame A with 200 rows and 5 columns and B with 6963 rows and 5 columns:
}
\usage{
megaplots_demo_data
}
\description{
List of datasets A and B containing example data
}
\keyword{internal}
