% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers_create_file.R
\name{read_dataset}
\alias{read_dataset}
\title{Helper function to read datasets}
\usage{
read_dataset(path)
}
\arguments{
\item{path}{Path to a file (.sas7bdat, .csv, .rdata). For `.RData`, the
first object listed by [load()] is returned (a warning is emitted if more
than one object exists).}
}
\description{
Helper function to read datasets
}
