% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_megaplot_data.R
\name{summarise_megaplot_data}
\alias{summarise_megaplot_data}
\title{Summarise Megaplot Data for Summary Panel}
\usage{
summarise_megaplot_data(data)
}
\arguments{
\item{data}{list of data sets used in megaplots}
}
\description{
Summarise Megaplot Data for Summary Panel
}
