% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_createfile_events.R
\name{createFile.events}
\alias{createFile.events}
\title{Function to create event level megaplots dataset from ADaM datasets}
\usage{
createFile.events(
  mp_data,
  path_data,
  id = "USUBJID",
  data_filter = NULL,
  param = list(c("AEBODSYS", "AEDECOD"), c("AEBODSYS", "AELLT"), c("AEDECOD", "AESEV"),
    c("AEDECOD", "AESER")),
  prefix = NULL,
  event_start = c("ASTDT", "AESTDT", "ADY"),
  event_end = c("AENDT", "AEENDT"),
  calc_time_to_first = FALSE,
  calc_days_with = FALSE,
  left_censor = NULL,
  keep_vars = NULL
)
}
\arguments{
\item{mp_data}{A list containing the subject-level and event data. It must have entries "sl" and "events" (may be NULL).}

\item{path_data}{A data frame or a file path to the dataset (SAS, CSV, or RData) to be read. File should be ADaM conform.}

\item{id}{A string representing the subject identifier column name. Default is "USUBJID".}

\item{data_filter}{A string for filtering the dataset using dplyr syntax. Default is NULL.}

\item{param}{A list of character vectors specifying the event group and event names for analysis.}

\item{prefix}{A list of character vectors for prefixes to be added to event group and event names.}

\item{event_start}{A character vector of possible event start date/time column names.}

\item{event_end}{A character vector of possible event end date/time column names.}

\item{calc_time_to_first}{A logical indicating whether to calculate the time to the first event. Default is FALSE.}

\item{calc_days_with}{A logical indicating whether to calculate the days with the event. Default is FALSE.}

\item{left_censor}{A numeric value specifying the left censoring time. Default NULL means no left cesnoring will be performed.}

\item{keep_vars}{A character vector of additional event level variables to keep in the final dataset.}
}
\value{
A modified list containing updated "sl" and "events" entries with the processed event-level data.
}
\description{
This function processes ADaM datasets to create a comprehensive dataset for event-level analysis with megaplots,
including the ability to calculate time to first event or days with event, left censor data and manage additional variables.
}
