% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_createfile_sl.R
\name{createFile.sl}
\alias{createFile.sl}
\title{Function to create a subject-level megaplots dataset from an ADSL (ADaM) dataset}
\usage{
createFile.sl(
  path_adsl,
  id = "USUBJID",
  data_filter = NULL,
  display_start_date = c("REFSTDT", "RFSTDT", "RFSTDTC", "RFICDT", "RANDDT", "TRTSTDT"),
  display_end_date = c("REFENDT", "RFENDTC", "RFENDT", "LVDT", "WDICDT"),
  relative_day_1 = c("TRTSTDT", "TRTSDT"),
  trtstdt = NULL,
  trtendt = NULL,
  trt = NULL
)
}
\arguments{
\item{path_adsl}{Path to the adsl-dataset (can be a dataframe or file path).}

\item{id}{Unique subject identifier (default: "USUBJID").}

\item{data_filter}{Subset dataset according to the filter conditions (default: NULL). Using dplyr::filter() syntax. Conditions should be wrapped in '' and concatenated by c(). Example: data_filter='SAFFL == "Y"'}

\item{display_start_date}{Columns to be used as start dates (date of the first contact to the participant). If a vector with multiple column names is given, the first one present in the provided dataset is used. Per default, the following column names are checked: "REFSTDT","RFSTDT","RFSTDTC","RFICDT","RANDDT","TRTSTDT"}

\item{display_end_date}{Columns to be used as end dates (date of the last contact to the participant. If a vector with multiple column names is given, the maximum of all present in the provided dataset is used. Per default, the following column names are checked: "REFENDT","RFENDTC","RFENDT", "LVDT", "WDICDT"}

\item{relative_day_1}{Columns to be used as the reference for relative day 1.}

\item{trtstdt}{Columns to be used as treatment start dates.}

\item{trtendt}{Columns to be used as treatment end dates.}

\item{trt}{Logical indicating whether to include treatment dates (default: TRUE).}
}
\value{
A list containing the processed subject level dataset and NULL for events.
}
\description{
This function reads an ADSL dataset from various file formats (SAS, CSV, RData) and processes it
to create a subject-level dataset suitable for generating megaplots. It allows for filtering of
the dataset based on specified conditions and computes relevant date variables.
}
