% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mega_plot.R
\name{draw_mega_plot}
\alias{draw_mega_plot}
\title{Title}
\usage{
draw_mega_plot(
  megaplot_prepared_data = megaplot_prepared_data(),
  megaplot_filtered_data = megaplot_filtered_data(),
  select_grouping = NULL,
  line_width = 3
)
}
\arguments{
\item{line_width}{}
}
\description{
Title
}
