% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mega_plot.R
\name{draw_mega_plot}
\alias{draw_mega_plot}
\title{Draws a 'Megaplot' graph}
\usage{
draw_mega_plot(
  megaplot_prepared_data = megaplot_prepared_data(),
  megaplot_filtered_data = megaplot_filtered_data(),
  select_grouping = NULL,
  line_width = 3,
  line_width_subjects,
  event_tooltips = TRUE
)
}
\arguments{
\item{megaplot_prepared_data}{data.frame with subject information used for the subject lines}

\item{megaplot_filtered_data}{data.frame with selected events for the event lines}

\item{select_grouping}{character vector with grouping variable names}

\item{line_width}{numeric value for event line width}

\item{line_width_subjects}{numeric value for subject line width}

\item{event_tooltips}{logical value if event tooltips should be turned on/off}
}
\description{
Draws a 'Megaplot' graph
}
