% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mp_finalize.R
\name{finalize_mp_object}
\alias{finalize_mp_object}
\title{Finalize Megaplots data for the app}
\usage{
finalize_mp_object(mp)
}
\arguments{
\item{mp}{A populated `mp_data_builder`.}
}
\value{
A data frame ready for use in the Megaplots app
}
\description{
Completes the build started with [init_mp_object()] and [add.events()] (and
optionally [add.sl_data()] if you attach subject-level data before events).
}
