% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Megaplots-package.R
\docType{package}
\name{Megaplots-package}
\alias{Megaplots}
\alias{Megaplots-package}
\title{Megaplots: A 'shiny' Application for Pattern Detection}
\description{
Depicts individual study courses per day. Pre-specified events on certain days, for example drug intake, efficacy, or safety events, are indicated by colored lines. The detection of patterns in the data or the temporal connection of study procedures and outcome events is supported by several functions, such as sorting, grouping, and changing the complexity of the graphical display.
}
\author{
\strong{Maintainer}: Steffen Jeske \email{steffen.jeske.ext@bayer.com}

}
\keyword{internal}
