% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers_create_file.R
\name{read_dataset}
\alias{read_dataset}
\title{Helper function to read datasets}
\usage{
read_dataset(path)
}
\arguments{
\item{path}{Path to a file (.sas7bdat, .csv, .rdata). If .rdata-format the file should best contain only one dataset.}
}
\description{
Helper function to read datasets
}
